Don't invoke the linestr callback if we found a keyword and the bufptr still contains an arrow at its beginning.
This means we're about to compile a method invocation, the tokenizer forced
parsing a word, and we got the const op of that.

I'm not too happy with this. I see how it can be useful to hook into those
places, but it should probably be a different feature of DD.